+++ src/games/Game.hpp
@@ -2,7 +2,6 @@
 class Game : public geode::Popup<> {
 protected:
     std::string m_name;
-    cocos2d::CCMenu* m_buttonMenu;
 
     bool initGame(std::string name);
 public: